+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix import of resolver mappings and order configuration for older relANNIS
+  versions.
+
 ## [2.0.2] - 2022-03-31
 
 ### Fixed